feat(easteregg/data.go): add new positive and rude phrases to CheckinPhrases and CheckoutPhrases to provide more variety and enhance user experience
The commit adds new phrases to the `CheckinPhrases` and `CheckoutPhrases` arrays in the `data.go` file of the `easteregg` package. These phrases are used to generate random messages when checking in and checking out. The new phrases aim to provide more variety and enhance the user experience by adding positive and rude phrases.